backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
celery[redis]==5.4.0
redis==5.0.6
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
scikit-learn==1.5.0
radon==6.0.1
prometheus-client==0.20.0
python-multipart==0.0.9
email-validator==2.1.1
numpy>=1.26.0
